上传页面布局修改，显示滚动条

+++ collection-console/src/main/java/com/keem/kochiu/collection/data/dto/ResourceDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ResourceDto {
 
     private int userId;
-    private int cateId;
+    private long cateId;
     private String sourceFileName;
     private String resourceUrl;
     private String fileExt;
